Update .eslintrc
Please lets apply some more rules I know it is a lot but there are files like DecodedBitStreamParser
that run completly out of scope for example they use switch functions that assign const outside of the switch statement but they do not depend on that behavior and that is well

https://eslint.org/docs/rules/no-case-declarations

i have a fork of this repo with a ESM build and i would not like to patch my Fork directly with all this needed stuff it would be much greater if we agree on that and i can start issuing a pr to this Repo do you agree? 

https://github.com/frank-dspeed/esnext-zxing-library/issues/1

shows the result of this change.